requirements.txt
webdriver-manager
allure-python-commons
allure-pytest
selenium
pytest
